abstract merge mode key handling into a hook instead of inlining into SceneView
