+++ src/components/SceneView.tsx
@@ -31,7 +31,11 @@ import SceneOptionsContext from '@/contexts/SceneOptionsContext';
 import { Box, IconButton, Tooltip, useTheme } from '@mui/material';
 import CenterFocusStrongIcon from '@mui/icons-material/CenterFocusStrong';
 import { SceneContextSetup } from '@/contexts/SceneContext';
-import { useLassoPath, useSceneTextureMapCache } from '@/hooks';
+import {
+  useLassoPath,
+  useSceneTextureMapCache,
+  useSelectionMergeModeKeys
+} from '@/hooks';
 import {
   EffectComposer,
   Outline,
@@ -75,8 +79,6 @@ export default function SceneView({ vertexInteractionMode }: SceneViewProps) {
     points: InteractionPoint[];
     selectionMergeMode: NodeSelectionMergeMode;
   }>();
-  const [isAltPressed, setIsAltPressed] = useState(false);
-  const [isShiftPressed, setIsShiftPressed] = useState(false);
   const [isScenePointerInside, setIsScenePointerInside] = useState(false);
   const canvasRef = useRef<HTMLCanvasElement>(null);
   const selectionMergeIndicatorRef = useRef<HTMLDivElement>(null);
@@ -120,51 +122,9 @@ export default function SceneView({ vertexInteractionMode }: SceneViewProps) {
   const textureCacheMap = useSceneTextureMapCache(textureDefs);
   const model = useAppSelector(selectModel);
   const theme = useTheme();
-
-  useEffect(() => {
-    const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === 'Alt') {
-        if (isScenePointerInsideRef.current) {
-          event.preventDefault();
-        }
-
-        setIsAltPressed(true);
-      }
-
-      if (event.key === 'Shift') {
-        setIsShiftPressed(true);
-      }
-    };
-
-    const handleKeyUp = (event: KeyboardEvent) => {
-      if (event.key === 'Alt') {
-        if (isScenePointerInsideRef.current) {
-          event.preventDefault();
-        }
-
-        setIsAltPressed(false);
-      }
-
-      if (event.key === 'Shift') {
-        setIsShiftPressed(false);
-      }
-    };
-
-    const handleWindowBlur = () => {
-      setIsAltPressed(false);
-      setIsShiftPressed(false);
-    };
-
-    window.addEventListener('keydown', handleKeyDown);
-    window.addEventListener('keyup', handleKeyUp);
-    window.addEventListener('blur', handleWindowBlur);
-
-    return () => {
-      window.removeEventListener('keydown', handleKeyDown);
-      window.removeEventListener('keyup', handleKeyUp);
-      window.removeEventListener('blur', handleWindowBlur);
-    };
-  }, []);
+  const { isAltPressed, isShiftPressed } = useSelectionMergeModeKeys(
+    isScenePointerInsideRef
+  );
 
   const canvasStyle = useMemo(() => {
     let cursor = 'default';
